+++ tests/test_entities.py
@@ -48,6 +48,7 @@ async def test_account_sensors(
         DOMAIN,
         f"{entry.entry_id}_account_last_updated",
     )
+    last_updated_entity = entity_registry.async_get(last_updated_entity_id)
     highest_level_entity_id = entity_registry.async_get_entity_id(
         "sensor",
         DOMAIN,
@@ -102,13 +103,8 @@ async def test_account_sensors(
         "Oak Logs",
         "Copper Ore",
     ]
-    last_updated_state = hass.states.get(last_updated_entity_id)
-    assert last_updated_state.state == "2026-06-29T12:00:00+00:00"
-    assert (
-        last_updated_state.attributes["source_updated_at"]
-        == "2026-06-29T12:00:00+00:00"
-    )
-    assert "last_successful_update" in last_updated_state.attributes
+    assert last_updated_entity.disabled_by is er.RegistryEntryDisabler.INTEGRATION
+    assert hass.states.get(last_updated_entity_id) is None
 
 
 async def test_character_sensors(